moved adapter interfaces from domain to app
