+++ app/adapters/DBAdapterInterface.go
@@ -1,0 +1,8 @@
+package adapters
+
+import "github.com/kosatnkn/db"
+
+// DBAdapterInterface is implemented by all database adapters.
+type DBAdapterInterface interface {
+	db.AdapterInterface
+}
