+++ app/container/Container.go
@@ -1,7 +1,7 @@
 package container
 
 import (
-	"github.com/kosatnkn/catalyst/domain/boundary/adapters"
+	"github.com/kosatnkn/catalyst/app/adapters"
 	"github.com/kosatnkn/catalyst/domain/boundary/repositories"
 	"github.com/kosatnkn/catalyst/domain/boundary/services"
 )
