+++ app/container/resolver.go
@@ -18,7 +18,7 @@ func Resolve(cfg *config.Config) *Container {
 	return &Container{
 		Adapters:     resolveAdapters(cfg),
 		Repositories: resolveRepositories(),
-		Services:     resolveServices(cfg.ServicesConfig),
+		Services:     resolveServices(cfg.Services),
 	}
 }
