+++ channels/http/controllers/Controller.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"net/http"
 
+	"github.com/kosatnkn/catalyst/app/adapters"
 	"github.com/kosatnkn/catalyst/app/container"
 	"github.com/kosatnkn/catalyst/channels/http/response"
-	"github.com/kosatnkn/catalyst/domain/boundary/adapters"
 	"github.com/kosatnkn/catalyst/domain/globals"
 )
