+++ channels/http/error/handler.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"net/http"
 
+	"github.com/kosatnkn/catalyst/app/adapters"
 	baseErrs "github.com/kosatnkn/catalyst/app/errors"
 	httpErrs "github.com/kosatnkn/catalyst/channels/http/errors"
-	"github.com/kosatnkn/catalyst/domain/boundary/adapters"
 	domainErrs "github.com/kosatnkn/catalyst/domain/errors"
 	externalErrs "github.com/kosatnkn/catalyst/externals/errors"
 )
