+++ domain/usecases/sample/Sample.go
@@ -1,8 +1,8 @@
 package sample
 
 import (
+	"github.com/kosatnkn/catalyst/app/adapters"
 	"github.com/kosatnkn/catalyst/app/container"
-	"github.com/kosatnkn/catalyst/domain/boundary/adapters"
 	"github.com/kosatnkn/catalyst/domain/boundary/repositories"
 )
