+++ externals/repositories/SampleMySQLRepository.go
@@ -3,7 +3,7 @@ package repositories
 import (
 	"context"
 
-	"github.com/kosatnkn/catalyst/domain/boundary/adapters"
+	"github.com/kosatnkn/catalyst/app/adapters"
 	"github.com/kosatnkn/catalyst/domain/boundary/repositories"
 	"github.com/kosatnkn/catalyst/domain/entities"
 )
