+++ go.mod
@@ -10,8 +10,9 @@ require (
 	github.com/google/uuid v1.1.2
 	github.com/gorilla/mux v1.8.0
 	github.com/iancoleman/strcase v0.1.2
+	github.com/kosatnkn/db v0.0.0-20210210053948-356124926aaf
 	github.com/kr/pretty v0.2.1 // indirect
-	github.com/lib/pq v1.8.0
+	github.com/lib/pq v1.9.0
 	github.com/logrusorgru/aurora v2.0.3+incompatible
 	github.com/prometheus/client_golang v1.8.0
 	golang.org/x/sys v0.0.0-20201017003518-b09fb700fbb7 // indirect
